Implement includes array as well as excludes.
In the project.json, you can now specify an includes array which will override
the excludes. This allows you to exclude a whole directory, but include a single
file in it.

Includes and excludes now apply to files as well as directories. This will allow
you to include a whole directory, but exclude a particular file.