Straighten out the new vs. modern translations
This is for strings where the translation was assembled from
the older translations. When there is a segment that is missing
and we don't have a translation for it, then the whole thing
goes into the new resources. If the whole thing is translated,
it goes into the modern resources.