Solved the case of the missing plurals
My dear Watson pointed me to the ruby file type which was sorely
lacking in changes related to the determination of new strings,
especially those of the plural persuasion. And so it was that I
found myself staring at the method that checks that all the plurals
were fully accounted. In my reveries, I discovered an important
clue. The parameters were sourceRes and res which to the lesser
mind would be perfectly innocuous. However, my intuition spoke to
me that such poorly named parametes were ripe for abuse. However
better named they should have been, they were used as the extracted
resource containing the source strings and the target resource
containing the translations. Following the clue, it lead me further
afield until I came upon the callers of said function, and I gazed
upon its inpenetrable coutenance and saw the culprit staring back
at me in cold bits. The arguments to the method were incorrect and
caused the death of the new resource calculation. I quickly grabbed
my trusty editor and captured the contempible arguments and delivered
them safely to the great bit bucket in the sky.  Later, I realized
the checking method could be cleaned up and made more clear, which
I did with such supreme vigour that the lady of the house looked
upon me with great amusement. I ran the tool again upon its intended
project and lo and behold, the new resources were correctly tallied
and can be dispatched to the translators with full expendiency.
And so was solved the mystery of the missing plural resources.