Fix markdown crash when the translation has more components than the source
The remark stringifier could not deal with a component node
type so it threw an exception when it was encountered. Instead, we delete
any component nodes in the translation that do not have a mapping
to a source component and pretend they were not there. It also gives
a warning.