Fix incorrect previous checkin
Previous checkin did not work properly because it avoided
writing extracted and new strings every time. This happened
because it checked if there were any strings in the extracted
translation set before any strings were added to that set.
The answer was always "no" of course! This fix checks the
number of strings in the translation set after all strings
have had the chance to be added to the set.